+++ flow-typed/common/timeline.js
@@ -2,3 +2,8 @@ declare type TimelineRow = {
   columns: Array<Array<string>>,
   date: string,
 };
+
+declare type TimelineGraphRow = {
+  date: string,
+  content: Node<*>
+};