+++ src/timeline-graph/index.js
@@ -1,2 +1,3 @@
 // @flow
-export { TimelineGraph } from "./timeline-graph";
+export { TimelineGraph } from "./timeline-graph-DEPRECATED/timeline-graph";
+export { TimelineGraph as CommonTimelineGraph } from "./timeline-graph";