+++ src/timeline-graph/timeline-graph.spec.js
@@ -2,132 +2,58 @@
 import "jest-dom/extend-expect";
 import React from "react";
 import moment from "moment-timezone";
-import { TestWrapper } from "../utils";
-import { TimelineGraph } from "./index";
-import { fireEvent, render } from "react-testing-library";
+import { CommonTimelineGraph } from "./index";
+import { TestWrapper } from "../utils/index";
+import { bindElementToQueries } from "dom-testing-library";
+import { cleanup, fireEvent, render } from "react-testing-library";
 
+afterEach(cleanup);
 moment.tz.setDefault("America/Los_Angeles");
+const bodyUtils = bindElementToQueries(document.body);
 
-test("render timeline graph without date", () => {
+const rows = [
+  {
+    date: "2018-04-20T16:20:00Z",
+    content: <div>Testing</div>,
+  },
+];
+
+test("render simple timeline graph", () => {
   const { container } = render(
     <TestWrapper>
-      <TimelineGraph
-        rows={[
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Test column"]],
-          },
-        ]}
-      />
+      <CommonTimelineGraph rows={rows} />
     </TestWrapper>,
   );
   expect(container).toMatchSnapshot();
 });
 
-test("render timeline graph with date", () => {
+test("test selecting an item on the timeline graph", () => {
+  const mockOnSelect = jest.fn(result => result);
   const { container } = render(
     <TestWrapper>
-      <TimelineGraph
-        rows={[
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Test column"]],
-          },
-        ]}
-        dateFormat="YYYY-MM-DD"
+      <CommonTimelineGraph
+        onSelect={mockOnSelect}
+        rows={rows}
       />
     </TestWrapper>,
   );
+  fireEvent.click(bodyUtils.getByTestId("timeline-item-0"));
+  expect(mockOnSelect.mock.results[0].value).toEqual(0);
   expect(container).toMatchSnapshot();
 });
 
-test("render and select timeline graph with multiple rows", () => {
-  const onSelectRow = jest.fn();
+test("test unselecting an item on the timeline graph", () => {
+  const mockOnSelect = jest.fn(result => result);
   const { container } = render(
     <TestWrapper>
-      <TimelineGraph
-        rows={[
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Row 0, Column 0, Field 0"]],
-          },
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Row 1, Column 0, Field 0"], ["Row 1, Column 1, Field 0", "Row 1, Column 1, Field 1"]],
-          },
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Row 2, Column 0, Field 0"]],
-          },
-        ]}
-        dateFormat="YYYY-MM-DD"
-        onSelectRow={onSelectRow}
-        selectedRowIndex={2}
+      <CommonTimelineGraph
+        onSelect={mockOnSelect}
+        selectedItem={0}
+        rows={rows}
       />
     </TestWrapper>,
   );
-  expect(container).toMatchSnapshot();
-  [1, 2, 0].forEach(rowIndex => {
-    const selectableRow = container.querySelector(`[data-row-index='${rowIndex}'] .card-container`);
-    if (selectableRow) {
-      onSelectRow.mockReset();
-      fireEvent.click(selectableRow);
-      expect(onSelectRow).toBeCalledWith(rowIndex);
-    } else {
-      expect(false).toBeTruthy();
-    }
-  });
-});
-
-test("render timeline graph with multiple columns", () => {
-  const { container } = render(
-    <TestWrapper>
-      <TimelineGraph
-        rows={[
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Column 0, Field 0"], ["Column 1, Field 0", "Column 1, Field 1"]],
-          },
-        ]}
-        dateFormat="YYYY-MM-DD"
-      />
-    </TestWrapper>,
-  );
-  expect(container).toMatchSnapshot();
-});
-
-test("render timeline graph with narrow cards", () => {
-  const { container } = render(
-    <TestWrapper>
-      <TimelineGraph
-        rows={[
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Test column"]],
-          },
-        ]}
-        dateFormat="YYYY-MM-DD"
-        cardWidth="narrow"
-      />
-    </TestWrapper>,
-  );
-  expect(container).toMatchSnapshot();
-});
-
-test("render timeline graph with tall cards", () => {
-  const { container } = render(
-    <TestWrapper>
-      <TimelineGraph
-        rows={[
-          {
-            date: "2018-04-20T16:20:00Z",
-            columns: [["Test column"]],
-          },
-        ]}
-        dateFormat="YYYY-MM-DD"
-        cardHeight="tall"
-      />
-    </TestWrapper>,
-  );
+  fireEvent.click(bodyUtils.getByTestId("timeline-item-0"));
+  expect(mockOnSelect.mock.results[0].value).toEqual(null);
   expect(container).toMatchSnapshot();
 });