+++ src/utils/index.js
@@ -1,4 +1,5 @@
 // @flow
-export { COUNTRIES, SERGEY } from "./constants";
+export { COUNTRIES, EXAMPLE_TABLE_DATA, SERGEY } from "./constants";
 export { ExampleBlock } from "./example-block";
 export { TestWrapper } from "./test-wrapper";
+export { StyleWrapper as ExampleWrapper } from "../style-wrapper";